requirements.txt
streamlit
requests
python-dotenv
pillow
opencv-python-headless
